demos-trainer/requirements.txt
# Full requirements to use demos-scraper training functions
# Including fat dependencies like Tensorflow and Keras

Automat==0.7.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0
PyDispatcher==2.0.5
PyHamcrest==1.9.0
Scrapy==1.5.1
Twisted==18.7.0
Unidecode==1.0.23
Werkzeug==0.14.1
absl-py==0.5.0
argparse==1.4.0
asn1crypto==0.24.0
astor==0.7.1
attrs==18.2.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
coverage==4.5.2
cryptography==2.3.1
cssselect==1.0.3
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
linecache2==1.0.0
lxml==4.2.5
matplotlib==3.0.3
mock==2.0.0
numpy==1.14.5
parsel==1.5.0
pbr==5.1.1
pdfminer.six==20170720
pip==18.1
protobuf==3.6.1
pyOpenSSL==18.0.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pycryptodome==3.6.6
queuelib==1.5.0
redis==3.0.1
requests==2.20.1
scikit-learn==0.20.0
scipy==1.1.0
service-identity==17.0.0
setuptools==39.1.0
six==1.11.0
tensorboard==1.13.1
tensorflow==1.13.1
termcolor==1.1.0
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.24.2
w3lib==1.19.0
wheel==0.31.1
yaml-config-parser==1.0.4
zope.interface==4.5.0